+++ e2e-tests/consts.ts
@@ -1,8 +1,8 @@
 // Copyright 2023 Parity Technologies (UK) Ltd.
 
 export const KUSAMA_ASSET_HUB_WS_URL = 'ws://127.0.0.1:9911';
-export const ROCOCO_ALICE_WS_URL = 'ws://127.0.0.1:9900';
-export const TRAPPIST_WS_URL = 'ws://127.0.0.1:9921';
+export const PASEO_ALICE_WS_URL = 'ws://127.0.0.1:9900';
+export const MUSE_WS_URL = 'ws://127.0.0.1:9921';
 
 export const BOB_ADDR = '5FHneW46xGXgs5mUiveU4sbTyGBzmstUspZC92UhjJM694ty';
 export const FERDE_ADDR = '5CiPPseXPECbkjWCa6MnjNokrgYjMqmKndv2rSnekmSK2DjL';