+++ examples/paseo/assetHub/foreignAssetTransfers/reserve/assetHubMUSEToBifrostRemoteReserve.ts
@@ -9,33 +9,38 @@ import { GREEN, PURPLE, RESET } from '../../../../colors';
 
 /**
  * In this example we are creating a `polkadotXcm` pallet `transferAssetsUsingTypeAndThen` call to send MUSE (foreign asset with location `{"parents":"1","interior":{"X1":{"Parachain":"3369"}}}`)
- * from a Rococo Asset Hub (System Parachain) account
- * to a Rhala Testnet (ParaChain) account, where the `xcmVersion` is set to safeXcmVersion, and there is no
+ * from a Paseo Asset Hub (System Parachain) account
+ * to a Bifrost Paseo (Parachain) testnet account, where the `xcmVersion` is set to safeXcmVersion, and there is no
  * `weightLimit` option provided which declares that the tx will allow unlimited weight to be used for fees.
- *  * The `paysWithFeeDest` value is set to pay fees with `MUSE` and the values for `assetTransferType` and `feesTransferType`
- * are both set to `LocalReserve` specifying that the reserve location to be used for transfers and fees is AssetHub.
+ * We provide a `paysWithFeeDest` asset location for MUSE which will be used to pay for fees on the destination chain.
+ * We also provide a value of `RemoteReserve` for the `assetTransferType` and `feeTransferType` options declaring a custom reserve location for both
+ * the `remoteReserveAssetTransferTypeLocation` and `remoteReserveFeesTransferTypeLocation` options.
  *
  * NOTE: To specify the amount of weight for the tx to use provide a `weightLimit` option containing desired values for `refTime` and `proofSize`.
  */
 const main = async () => {
-	const { api, specName, safeXcmVersion, chainName } = await constructApiPromise(
-		'wss://rococo-asset-hub-rpc.polkadot.io',
+	const { api, specName, chainName, safeXcmVersion } = await constructApiPromise(
+		'wss://paseo-asset-hub-rpc.polkadot.io',
 	);
-	const assetApi = new AssetTransferApi(api, specName, safeXcmVersion, { chainName });
+	const assetApi = new AssetTransferApi(api, specName, safeXcmVersion, {
+		chainName,
+	});
 
 	let callInfo: TxResult<'call'>;
 	try {
 		callInfo = await assetApi.createTransferTransaction(
-			'2004',
-			'5EWNeodpcQ6iYibJ3jmWVe85nsok1EDG8Kk3aFg8ZzpfY1qX',
+			`2030`,
+			'0x6E733286C3Dc52C67b8DAdFDd634eD9c3Fb05B5B',
 			[`{"parents":"1","interior":{"X1":{"Parachain":"3369"}}}`],
 			['1000000000000'],
 			{
 				format: 'call',
 				xcmVersion: safeXcmVersion,
 				paysWithFeeDest: '{"parents":"1","interior":{"X1":{"Parachain":"3369"}}}',
-				assetTransferType: 'LocalReserve',
-				feesTransferType: 'LocalReserve',
+				assetTransferType: 'RemoteReserve',
+				remoteReserveAssetTransferTypeLocation: '{"parents":"1","interior":{"X1":{"Parachain":"3369"}}}',
+				feesTransferType: 'RemoteReserve',
+				remoteReserveFeesTransferTypeLocation: '{"parents":"1","interior":{"X1":{"Parachain":"3369"}}}',
 			},
 		);