+++ examples/paseo/assetHub/foreignAssetTransfers/reserve/assetHubWETHToHydration.ts
@@ -9,8 +9,8 @@ import { GREEN, PURPLE, RESET } from '../../../../colors';
 
 /**
  * In this example we are creating a `polkadotXcm` pallet `transferAssetsUsingTypeAndThen` call to send WETH (foreign asset with location `{"parents":"2","interior":{"X2":[{"GlobalConsensus":{"Ethereum":{"chainId":"11155111"}}},{"AccountKey20":{"network":null,"key":"0xfff9976782d46cc05630d1f6ebab18b2324d6b14"}}]}}`)
- * from a Rococo Asset Hub (System Parachain) account
- * to a Rhala Testnet (ParaChain) account, where the `xcmVersion` is set to safeXcmVersion, and there is no
+ * from a Paseo Asset Hub (System Parachain) account
+ * to a Hydration Paseo (ParaChain) account, where the `xcmVersion` is set to safeXcmVersion, and there is no
  * `weightLimit` option provided which declares that the tx will allow unlimited weight to be used for fees.
  * The `paysWithFeeDest` value is set to pay fees with WETH and the values for `assetTransferType` and `feesTransferType`
  * are both set to `LocalReserve` specifying that the reserve location to be used for transfers and fees is AssetHub.
@@ -19,14 +19,14 @@ import { GREEN, PURPLE, RESET } from '../../../../colors';
  */
 const main = async () => {
 	const { api, specName, safeXcmVersion, chainName } = await constructApiPromise(
-		'wss://rococo-asset-hub-rpc.polkadot.io',
+		'wss://paseo-asset-hub-rpc.polkadot.io',
 	);
 	const assetApi = new AssetTransferApi(api, specName, safeXcmVersion, { chainName });
 
 	let callInfo: TxResult<'call'>;
 	try {
 		callInfo = await assetApi.createTransferTransaction(
-			'2004',
+			'2034',
 			'5EWNeodpcQ6iYibJ3jmWVe85nsok1EDG8Kk3aFg8ZzpfY1qX',
 			[
 				`{"parents":"2","interior":{"X2":[{"GlobalConsensus":{"Ethereum":{"chainId":"11155111"}}},{"AccountKey20":{"network":null,"key":"0xfff9976782d46cc05630d1f6ebab18b2324d6b14"}}]}}`,