+++ examples/polkadot/parachain/paysWithFeeDest/bifrostToAssetHubPaysWithDOT.ts
@@ -8,27 +8,30 @@ import { TxResult } from '../../../../src/types';
 import { GREEN, PURPLE, RESET } from '../../../colors';
 
 /**
- * In this example we are creating a `polkadotXcm` pallet `transferAssets` call to send 1 WND (foreign asset with location `{"parents":"2","interior":{"X1":{"GlobalConsensus":"Westend"}}}`)
- * from a Rococo Asset Hub (System Parachain) account
- * to a Westend Asset Hub account, where the `xcmVersion` is set to safeXcmVersion and no `weightLimit` option is provided declaring that
- * the tx will allow unlimited weight to be used for fees.
+ * In this example we are creating a `transferAssets` call to send WETH
+ * from a Bifrost Polkadot (Parachain) account
+ * to a Polkadot AssetHub (System Parachain) account, where the `xcmVersion` is set to 3 and no `weightLimit` is provided declaring that
+ * the allowable weight will be `unlimited` and `paysWithFeeDest` is asset ID `DOT` (Polkadot)
+ * declaring that `DOT` `should be used to pay for tx fees on the destination chain.
  *
  * NOTE: To specify the amount of weight for the tx to use provide a `weightLimit` option containing desired values for `refTime` and `proofSize`.
  */
 const main = async () => {
-	const { api, specName, safeXcmVersion } = await constructApiPromise('wss://rococo-asset-hub-rpc.polkadot.io');
-	const assetApi = new AssetTransferApi(api, specName, safeXcmVersion);
+	const xcmVersion = 3;
+	const { api, specName } = await constructApiPromise('wss://bifrost-polkadot-rpc.dwellir.com/ws');
+	const assetApi = new AssetTransferApi(api, specName, xcmVersion);
 
 	let callInfo: TxResult<'call'>;
 	try {
 		callInfo = await assetApi.createTransferTransaction(
-			`{"parents":"2","interior":{"X2":[{"GlobalConsensus":"Westend"},{"Parachain":"1000"}]}}`,
+			'1000',
 			'5EWNeodpcQ6iYibJ3jmWVe85nsok1EDG8Kk3aFg8ZzpfY1qX',
-			[`{"parents":"2","interior":{"X1":{"GlobalConsensus":"Westend"}}}`],
-			['1000000000000'],
+			['WETH', 'DOT'],
+			['1000000000000', '10000000000'],
 			{
 				format: 'call',
-				xcmVersion: safeXcmVersion,
+				xcmVersion,
+				paysWithFeeDest: 'DOT', // Asset to be used to pay for fees on destination chain
 			},
 		);