+++ src/errors/checkXcmTxInputs.spec.ts
@@ -629,40 +629,16 @@ describe('checkAssetIdsLengthIsValid', () => {
 	it('Should correctly error when more than 2 assetIds are passed in', () => {
 		const assetIds = ['ksm', '1984', '10'];
 		const assetTransferType = undefined;
-		const xcmPallet = XcmPalletName.polkadotXcm;
 
-		const err = () => checkAssetIdsLengthIsValid(assetIds, xcmPallet, assetTransferType);
+		const err = () => checkAssetIdsLengthIsValid(assetIds, assetTransferType);
 		expect(err).toThrow('Maximum number of assets allowed for transfer is 2. Found 3 assetIds');
 	});
 	it('Should correctly not error when 2 or fewer assetIds are passed in and assetTransferType is defined', () => {
 		const assetIds = ['ksm', '1984'];
 		const assetTransferType = 'RemoteReserve';
-		const xcmPallet = XcmPalletName.polkadotXcm;
 
-		const err = () => checkAssetIdsLengthIsValid(assetIds, xcmPallet, assetTransferType);
+		const err = () => checkAssetIdsLengthIsValid(assetIds, assetTransferType);
 		expect(err).not.toThrow('Maximum number of assets allowed for transfer is 2. Found 3 assetIds');
-	});
-	it('Should correctly throw an error when provided more than 1 asset id for a non `transferAssetsUsingTypeAndThen` `xcmPallet` call', () => {
-		const assetIds = [
-			`{"parents":"1","interior":{"Here":""}}`,
-			`{"parents":"2","interior":{"X1":{"GlobalConsensus":"Westend"}}}`,
-		];
-		const assetTransferType = undefined;
-		const xcmPallet = XcmPalletName.xcmPallet;
-
-		const err = () => checkAssetIdsLengthIsValid(assetIds, xcmPallet, assetTransferType);
-		expect(err).toThrow('transferAssets transactions cannot contain more than 1 asset location id. Found 2 assetIds');
-	});
-	it('Should correctly throw an error when provided more than 1 asset id for a non `transferAssetsUsingTypeAndThen` `polkadotXcm` call', () => {
-		const assetIds = [
-			`{"parents":"1","interior":{"Here":""}}`,
-			`{"parents":"2","interior":{"X1":{"GlobalConsensus":"Westend"}}}`,
-		];
-		const assetTransferType = undefined;
-		const xcmPallet = XcmPalletName.polkadotXcm;
-
-		const err = () => checkAssetIdsLengthIsValid(assetIds, xcmPallet, assetTransferType);
-		expect(err).toThrow('transferAssets transactions cannot contain more than 1 asset location id. Found 2 assetIds');
 	});
 	it('Should correctly not throw an error when provided more than 1 asset id for a `transferAssetsUsingTypeAndThen` call', () => {
 		const assetIds = [
@@ -670,9 +646,8 @@ describe('checkAssetIdsLengthIsValid', () => {
 			`{"parents":"2","interior":{"X1":{"GlobalConsensus":"Westend"}}}`,
 		];
 		const assetTransferType = 'RemoteReserve';
-		const xcmPallet = XcmPalletName.polkadotXcm;
 
-		const err = () => checkAssetIdsLengthIsValid(assetIds, xcmPallet, assetTransferType);
+		const err = () => checkAssetIdsLengthIsValid(assetIds, assetTransferType);
 		expect(err).not.toThrow();
 	});
 });
@@ -874,13 +849,13 @@ describe('checkBridgeTxInputs', () => {
 
 describe('checkPaysWithFeeDestAssetIdIsInAssets', () => {
 	it('Should correctly error when a paysWithFeeDestAsset is not found in the assetIds list', () => {
-		const assetIds = [`{"parents":"2","interior":{"X1":{"GlobalConsensus":"Rococo"}}}`];
+		const assetIds = [`{"parents":"2","interior":{"X1":{"GlobalConsensus":"Paseo"}}}`];
 		const paysWithFeeDest = `{"parents":"2","interior":{"X2":[{"GlobalConsensus":{"Ethereum":{"chainId":"11155111"}}},{"AccountKey20":{"network":null,"key":"0xfff9976782d46cc05630d1f6ebab18b2324d6b14"}}]}}`;
 
 		const err = () => checkPaysWithFeeDestAssetIdIsInAssets(assetIds, paysWithFeeDest);
 
 		expect(err).toThrow(
-			'paysWithFeeDest asset must be present in assets to be transferred. Did not find {"parents":"2","interior":{"X2":[{"GlobalConsensus":{"Ethereum":{"chainId":"11155111"}}},{"AccountKey20":{"network":null,"key":"0xfff9976782d46cc05630d1f6ebab18b2324d6b14"}}]}} in {"parents":"2","interior":{"X1":{"GlobalConsensus":"Rococo"}}}',
+			'paysWithFeeDest asset must be present in assets to be transferred. Did not find {"parents":"2","interior":{"X2":[{"GlobalConsensus":{"Ethereum":{"chainId":"11155111"}}},{"AccountKey20":{"network":null,"key":"0xfff9976782d46cc05630d1f6ebab18b2324d6b14"}}]}} in {"parents":"2","interior":{"X1":{"GlobalConsensus":"Paseo"}}}',
 		);
 	});
 });