+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.4.3](https://github.com/paritytech/asset-transfer-api/compare/v0.4.2..v0.4.3)
+
+### Fix
+
+- fix: enable paysWithFeeDest for transferAssets ([#444](https://github.com/paritytech/asset-transfer-api/pull/444))([2a78ecb](https://github.com/paritytech/asset-transfer-api/commit/63b70328bdca2a9b48594f11252383efa2a78ecb))
+
+## [0.4.2](https://github.com/paritytech/asset-transfer-api/compare/v0.4.1..v0.4.2)
+
+### Fix
+
+- fix: default to xcmPallet ([#442](https://github.com/paritytech/asset-transfer-api/pull/442))([771f264](https://github.com/paritytech/asset-transfer-api/commit/81904a2df0c5e9f03cc936e461fedc4f4771f264))
+
 ## [0.4.1](https://github.com/paritytech/asset-transfer-api/compare/v0.4.0..v0.4.1)
 
 ### Chore